pyproject.toml
[tool.poetry]
name = "landscape-charm"
description = "Juju charm for deploying and managing Landscape Server"
authors = ["Canonical"]
package-mode = false

[tool.poetry.dependencies]
charmlibs-interfaces-tls-certificates = "1.3.0"
cosl = "1.4.0"
jinja2 = "3.1.6"
ops = "2.23.1"
python = "^3.10"

[tool.poetry.group.dev.dependencies]
black = "~=24.10.0"
charmcraftcache = "0.8.0"
coverage = "7.13.1"
flake8 = "7.3.0"
grpcio = "1.78.0"
isort = "7.0.0"
ops = { extras = ["testing"], version = "2.23.1" }
pydantic = "1.10.26"
pytest = "9.0.2"
ruff = "0.14.11"
urllib3 = "2.6.3"

[tool.poetry.group.integration.dependencies]
grpcio = "1.78.0"
jubilant = "1.6.2"
requests = "2.32.5"
urllib3 = "2.6.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ["py310"]
line-length = 88

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88
target-version = "py310"
# Skip vendored dependencies
extend-exclude = ["lib"]

[tool.ruff.lint]
select = [
  # pycodestyle errors
  "E",
  # pycodestyle warnings
  "W",
  # Pyflakes
  "F",
  # isort
  "I",
]

[tool.ruff.lint.isort]
combine-as-imports = true
force-sort-within-sections = true
order-by-type = false
case-sensitive = false

[tool.pytest.ini_options]
filterwarnings = ["once:.*Harness is deprecated.*:PendingDeprecationWarning"]
markers = [
    "disable_haproxy_mocks: disable automatic HAProxy mocking for tests that need real implementations"
]
pythonpath = [".", "lib", "src"]
testpaths = ["tests"]
